update
update

+++ init.py
@@ -1,10 +1,14 @@
 #!/usr/bin/env python
 #-*- coding:utf-8 -*-
 
+import os
+
 from app import db
 from app.models import User, Team, Submit
 
-db.drop_all()
+if os.path.exists(r'data\data.db'):
+    db.drop_all()
+
 db.create_all()
 db.session.add(User('rikka', 'yuta', True))
 db.session.commit()
